requirements.txt
pip
Pymoe
pillow
pymongo
dnspython
discord.py
asyncio
memory_profiler
